backend/requirements.txt
aniso8601==7.0.0
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
freeze==1.0.10
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
mysqlclient==1.4.4
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
promise==2.2.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
